many changes, but not 100% working! don't use now
editor:
bricks, floor, pipes: can't change width-property anymore, instead use the blocks one
show function/class constructor signature in title when using it
automatic backup file if the editor crashes
fixed design-bug
fix crash on set direction=None on turtles

game:
added levelpackswitcher if you install some level pack (no levelpack released yet)
delete points when your tux is dead
fixed flyingcircle levelscript
made game object speed's relative to frame rate

To show FPS, press 'f'

+++ littleRunner/Drawing/DrawGDI/Draw_GDI.cs
@@ -16,30 +16,30 @@ namespace littleRunner.Drawing.GDI
 
 
 
-        public override void DrawImage(dImage image, int x, int y, int width, int height)
+        public override void DrawImage(dImage image, float x, float y, int width, int height)
         {
             dImage_GDI gdiImage = (dImage_GDI)image;
-            g.DrawImage(gdiImage.ToGDI(), x, y, width, height);
+            g.DrawImage(gdiImage.ToGDI(), (int)x, (int)y, width, height);
         }
-        public override void DrawRectangle(dPen pen, int x, int y, int width, int height)
+        public override void DrawRectangle(dPen pen, float x, float y, int width, int height)
         {
-            g.DrawRectangle(pen.ToGDIPen(), x, y, width, height);
+            g.DrawRectangle(pen.ToGDIPen(), (int)x, (int)y, width, height);
         }
-        public override void FillRectangle(dPen pen, int x, int y, int width, int height)
+        public override void FillRectangle(dPen pen, float x, float y, int width, int height)
         {
-            g.FillRectangle(pen.ToGDIBrush(), x, y, width, height);
+            g.FillRectangle(pen.ToGDIBrush(), (int)x, (int)y, width, height);
         }
-        public override void DrawString(string text, dFont font, dColor color, int x, int y)
+        public override void DrawString(string text, dFont font, dColor color, float x, float y)
         {
             System.Drawing.SolidBrush b = new System.Drawing.SolidBrush(color.ToGDI());
             System.Drawing.Font f = new System.Drawing.Font(font.Family, font.Size, font.Style.ToGDI());
-            g.DrawString(text, f, b, x, y, font.Format.ToGDI());
+            g.DrawString(text, f, b, (int)x, (int)y, font.Format.ToGDI());
         }
 
 
-        public override void MoveCoords(int x, int y)
+        public override void MoveCoords(float x, float y)
         {
-            g.TranslateTransform(x, y);
+            g.TranslateTransform((int)x, (int)y);
         }
     }
 }
