+++ littleRunner/GameObjects/Enemies/Enemy.cs
@@ -29,9 +29,9 @@ namespace littleRunner.GameObjects
             set { startAtViewpoint = value; }
         }
 
-        virtual public void Check(out Dictionary<string, int> newpos)
+        virtual public void Check(out Dictionary<string, float> newpos)
         {
-            newpos = new Dictionary<string, int>();
+            newpos = new Dictionary<string, float>();
             newpos["top"] = 0;
             newpos["left"] = 0;
