+++ littleRunner/GameObjects/MainGameObjects/MainGameObject.cs
@@ -32,11 +32,11 @@ namespace littleRunner.GameObjects
         public abstract MainGameObjectMode Mode { get; set; }
 
 
-        public virtual Dictionary<string,int> Check(List<GameKey> pressedKeys)
+        public virtual Dictionary<string, float> Check(List<GameKey> pressedKeys)
         {
-            return new Dictionary<string, int>() { { "newtop", 0 }, { "newleft", 0 } };
+            return new Dictionary<string, float>() { { "newtop", 0 }, { "newleft", 0 } };
         }
-        public abstract void Move(MoveType mtype, int value, GameInstruction instruction);
+        public abstract void Move(MoveType mtype, float value, GameInstruction instruction);
         
         public virtual void getEvent(GameEvent gevent, Dictionary<GameEventArg, object> args)
         {
